Fix the weekly distance button
This commit fixes the weekly distance button and adds 3 precompiled entries to the training record when the program starts so that there's some data to play with once the program starts without the user having to manually fill the program in every time he runs it.
Before the weekly distance button would always return that the athlete swam, cycled, ran 0km in the output area no matter what the user inputted, it also didn't display the athletes records. This was a problem because it was not a correct output.
Now the weekly distance button returns the kilometres that the searched athlete ran/swam/cycled over the past week. It also displays all the individual training records of that athlete below the latter line in the output area. This is the desired result because it was the requested output in task 14 of the assignment.